Updated Status Page with Ajax updates and animations.
Replaced meta refresh (of 10 seconds) with an Ajax comparison of the previous status which occurs runs every 5 seconds. The entire page is now only reloaded when the status changes. The background now pulses a fades to a brighter version of the background color, then back when the page is loaded to indicate that status has changed.